feat: Phase 29 Receipts & Payments complete + Phase 27-28 gaps fixed (v4.10.4)
== Phase 29 — Receipts & Payments Module ==
NEW: Cheque entity + ChequeStatus enum + ChequeService (full CRUD + lifecycle)
NEW: DailyClosure domain entity (ClosingBalance/Difference computed in factory)
NEW: ChequeConfiguration + DailyClosureConfiguration (Fluent API, unique indexes)
NEW: Cheque DTOs, Requests, Validators, DI registration
FIX: CashBox.CategoryId field added (was missing from entity, existed in DTO)
FIX: CustomerReceiptService now uses IDocumentSequenceService (removed max+1)
FIX: SupplierPaymentService now uses GetNextIntAsync (was prefix-based)
FIX: CustomerReceiptViews code-behind double-VM-creation bug
ENH: Status badge color-coding in CustomerReceiptListView (Draft/Posted/Cancelled)

== Phase 27-28 Gaps Closed ==
FIX: SalesReturn.Post()/Cancel() timestamps (PostedAt/CancelledAt)
FIX: SalesReturnService injects IAccountingIntegrationService (journal entries)
FIX: ReverseSalesReturnEntryAsync added to AccountingIntegrationService
FIX: ProductUnitId=1 hardcoding removed from 6 ViewModels (8 occurrences)
FIX: AdditionalChargeAllocator extracted to standalone static helper
FIX: AllowBelowCostSale default true + warning-only (was blocking)
ENH: FlexibleInputCalculator helper for editable LineTotal column

== Accounting Integration ==
FIX: CreateSalesReturnEntryAsync + ReverseSalesReturnEntryAsync
FIX: Per-entity Customer.AccountId routing in return journal entries
NEW: SystemAccountKey.DeliveryChargesRevenue = 21
NEW: Account 1533 seeded for delivery charges under 1530 parent

== Documentation ==
UPDATED: AGENTS.md (RULE-517-520, FORBIDDEN, checklist)
UPDATED: README.md (v4.10.4+)
UPDATED: docs/CHANGELOG.md (v4.10.4 entry)
UPDATED: .opencode/agent/code-reviewer.md, implement-agent.md, orchestrator.md

== Build & Tests ==
Build: 0 errors, 0 warnings across all 11 projects
Tests: 1,600 passed, 71 skipped, 0 failed

+++ SalesSystem/SalesSystem.Api/Validators/AddProductUnitRequestValidator.cs
@@ -8,7 +8,7 @@ public class AddProductUnitRequestValidator : AbstractValidator<AddProductUnitRe
     public AddProductUnitRequestValidator()
     {
         RuleFor(x => x.UnitId)
-            .GreaterThan(0).WithMessage("يجب اختيار وحدة قياس صحيحة");
+            .GreaterThan((short)0).WithMessage("يجب اختيار وحدة قياس صحيحة");
 
         RuleFor(x => x.Factor)
             .GreaterThan(0).WithMessage("معامل التحويل يجب أن يكون أكبر من صفر");
