+++ SalesSystem/SalesSystem.Contracts/Requests/UpdateProductUnitRequest.cs
@@ -1,5 +1,5 @@
 namespace SalesSystem.Contracts.Requests;
 
 public record UpdateProductUnitRequest(
-    int UnitId
+    short UnitId
 );
