+++ SalesSystem/Tests/SalesSystem.Domain.Tests/Entities/ProductUnitTests.cs
@@ -9,7 +9,7 @@ public class ProductUnitTests
     [Fact]
     public void CreateBaseUnit_ShouldSetFactorToOne()
     {
-        var unit = ProductUnit.CreateBaseUnit(1, unitId: 1);
+        var unit = ProductUnit.CreateBaseUnit(1, unitId: (short)1);
         Assert.Equal(1, unit.Factor);
         Assert.True(unit.IsBaseUnit);
     }
@@ -17,22 +17,22 @@ public class ProductUnitTests
     [Fact]
     public void CreateBaseUnit_InvalidUnitId_ShouldThrowDomainException()
     {
-        Assert.Throws<DomainException>(() => ProductUnit.CreateBaseUnit(1, unitId: 0));
+        Assert.Throws<DomainException>(() => ProductUnit.CreateBaseUnit(1, unitId: (short)0));
     }
 
     [Fact]
     public void CreateDerivedUnit_FactorLessThanOrEqualToOne_ShouldThrow()
     {
         Assert.Throws<DomainException>(() =>
-            ProductUnit.CreateDerivedUnit(1, unitId: 2, factor: 1));
+            ProductUnit.CreateDerivedUnit(1, unitId: (short)2, factor: 1));
         Assert.Throws<DomainException>(() =>
-            ProductUnit.CreateDerivedUnit(1, unitId: 2, factor: 0.5m));
+            ProductUnit.CreateDerivedUnit(1, unitId: (short)2, factor: 0.5m));
     }
 
     [Fact]
     public void CreateDerivedUnit_FactorMoreThanOne_ShouldSucceed()
     {
-        var unit = ProductUnit.CreateDerivedUnit(1, unitId: 2, factor: 12);
+        var unit = ProductUnit.CreateDerivedUnit(1, unitId: (short)2, factor: 12);
         Assert.Equal(12, unit.Factor);
         Assert.False(unit.IsBaseUnit);
     }
@@ -40,8 +40,8 @@ public class ProductUnitTests
     [Fact]
     public void ToBaseUnitQuantity_ShouldMultiplyByFactor()
     {
-        var baseUnit = ProductUnit.CreateBaseUnit(1, unitId: 1);
-        var derivedUnit = ProductUnit.CreateDerivedUnit(1, unitId: 2, factor: 12);
+        var baseUnit = ProductUnit.CreateBaseUnit(1, unitId: (short)1);
+        var derivedUnit = ProductUnit.CreateDerivedUnit(1, unitId: (short)2, factor: 12);
 
         Assert.Equal(5, baseUnit.ToBaseUnitQuantity(5));
         Assert.Equal(60, derivedUnit.ToBaseUnitQuantity(5));
@@ -50,22 +50,22 @@ public class ProductUnitTests
     [Fact]
     public void CreateDerivedUnit_SortOrder_DefaultsToOne()
     {
-        var unit = ProductUnit.CreateDerivedUnit(1, unitId: 2, factor: 12);
+        var unit = ProductUnit.CreateDerivedUnit(1, unitId: (short)2, factor: 12);
         Assert.NotNull(unit);
     }
 
     [Fact]
     public void ChangeUnit_ValidUnitId_UpdatesUnitId()
     {
-        var unit = ProductUnit.CreateBaseUnit(1, unitId: 1);
-        unit.ChangeUnit(2);
+        var unit = ProductUnit.CreateBaseUnit(1, unitId: (short)1);
+        unit.ChangeUnit((short)2);
         Assert.Equal((short)2, unit.UnitId);
     }
 
     [Fact]
     public void ChangeUnit_InvalidUnitId_ThrowsDomainException()
     {
-        var unit = ProductUnit.CreateBaseUnit(1, unitId: 1);
+        var unit = ProductUnit.CreateBaseUnit(1, unitId: (short)1);
         Assert.Throws<DomainException>(() => unit.ChangeUnit(0));
     }
 }
